Moved rush-state::context::Context to rush-state::shell:Context.
