+++ rush-console/src/reader.rs
@@ -7,7 +7,7 @@ use crossterm::style::{Print, Stylize};
 use crossterm::terminal::{self, Clear, ClearType};
 use crossterm::{execute, queue};
 
-use rush_state::context::Context;
+use rush_state::shell::Context;
 
 // Represents a character that can be added to the line buffer, or an ENTER keypress, which will send the line buffer to the shell
 // Keypresses that may have been handled downstream, but should not result in any further behavior, are represented by the Ignored variant
@@ -142,7 +142,7 @@ fn generate_prompt(context: &Context) -> String {
     };
 
     // ? What is the actual name for this?
-    let prompt_tick = match context.command_success {
+    let prompt_tick = match context.success() {
         true => "❯".green(),
         false => "❯".red(),
     }
