+++ rush-state/src/lib.rs
@@ -1,5 +1,4 @@
 mod config;
-pub mod context;
 mod environment;
 pub mod errors;
 pub mod path;
